+++ __tests__/procedures/parser.test.ts
@@ -1,4 +1,22 @@
-import { temmeParser, TemmeSelector } from '../../src'
+import { DEFAULT_PROCEDURE_NAME, temmeParser, TemmeSelector } from '../../src'
+
+test('parse default procedure', () => {
+  const actual = temmeParser.parse(`div{ $foo };`)
+  const expected: TemmeSelector[] = [
+    {
+      type: 'normal-selector',
+      sections: [{ combinator: ' ', element: 'div', qualifiers: [] }],
+      procedure: {
+        name: DEFAULT_PROCEDURE_NAME,
+        args: [{ name: 'foo', filterList: [], modifier: null }],
+      },
+      arrayCapture: null,
+      children: [],
+    },
+  ]
+
+  expect(actual).toEqual(expected)
+})
 
 test('parse procedure', () => {
   const actual = temmeParser.parse(`div{ fn($c, 100) }`)
